+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.21.0] - 2023-10-13
+
 ### Added
 
 - Mutate Vintage Workload Cluster Apps to disable Pod Security Policies.
@@ -289,7 +291,8 @@ allow app CR creation.
 - Add validation webhook that is enabled for app CRs with
 `app-operator.giantswarm.io/version` label value >= `3.0.0`.
 
-[Unreleased]: https://github.com/giantswarm/app-admission-controller/compare/v0.20.0...HEAD
+[Unreleased]: https://github.com/giantswarm/app-admission-controller/compare/v0.21.0...HEAD
+[0.21.0]: https://github.com/giantswarm/app-admission-controller/compare/v0.20.0...v0.21.0
 [0.20.0]: https://github.com/giantswarm/app-admission-controller/compare/v0.19.0...v0.20.0
 [0.19.0]: https://github.com/giantswarm/app-admission-controller/compare/v0.18.7...v0.19.0
 [0.18.7]: https://github.com/giantswarm/app-admission-controller/compare/v0.18.6...v0.18.7